requirements.txt
torch~=1.10.0
pyyaml~=6.0
tqdm~=4.62.3
opencv-python~=4.5.4.60
matplotlib~=3.5.0
numpy~=1.21.2
pandas~=1.3.4
seaborn~=0.11.2
pillow~=8.4.0
torchvision~=0.11.1
requests~=2.26.0
thop~=0.0.31-2005241907
fastapi~=0.70.1
pydantic>=1.8.0,<2.0.0
uvicorn==0.16.0
python-multipart==0.0.5

